+++ chopin-typescript/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [typescript v0.1.0] chopin-typescript
 ### Releases
 - [NPM v0.1.0] https://www.npmjs.com/package/@chopinframework/sdk/v/0.1.0 - chopin-typescript
+
+## 2025-03-07 00:39:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.513.2 (2.545.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.1.1] chopin-typescript
+### Releases
+- [NPM v0.1.1] https://www.npmjs.com/package/@chopinframework/sdk/v/0.1.1 - chopin-typescript